feat: 레이아웃 수정 (#51)
